+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 Changelog
 =========
 
+Version 0.82.0
+--------------
+* added nkuluflag module
+
+Version 0.81.7
+--------------
+* bugfixes
+* added whisper feature extractor
+
 Version 0.81.6
 --------------
 * updated documentation